Added --local-only option to demo script for testing support

+++ demo.sh
@@ -1,6 +1,15 @@
 #!/bin/bash
 
-export INITIAL_GIT_TAG=v-1.0
+export INITIAL_GIT_TAG=demo-start
+
+push_changes=true
+
+# Parse arguments (keeping it simple for now)
+if [ "$1" == "--local-only" ]; then
+		echo "Running demo against local repo only..."
+		echo ""
+		push_changes=false
+fi
 
 function help() {
 	echo "Select from the following options: "
@@ -8,7 +17,7 @@ function help() {
 	printf " %s \t %s \n" "2" "Fix the bug"
 	printf " %s \t %s \n" "3" "Reset environment"
 	printf " %s \t %s \n" "4" "Exit"
-	
+
 	echo -e "Please enter your option: \c"
 	read OPTION
 
@@ -22,23 +31,25 @@ function help() {
 		add_feature_fix
 		;;
 	3 )
-		logInfo "Resetting the branch"		
-		reset				
+		logInfo "Resetting the branch"
+		reset
 		;;
 	4 )
-		logSuccess "Have a good day!"			
+		logSuccess "Have a good day!"
 		exit 0
 		;;
-	esac	
+	esac
 }
 
 function add_feature() {
 	cp -r code_staging/feature/* .
 	git add .
 	git commit -m "Adding the feature"
-	git push origin master
+	if [ "$push_changes" = true ]; then
+		git push origin master
+	fi
 	logSuccess "Feature added and committed"
-	
+
 	help
 }
 
@@ -46,23 +57,27 @@ function add_feature_fix() {
 	cp -r code_staging/fix/* .
 	git add .
 	git commit -m "Adding the feature fix"
-	git push origin master
+	if [ "$push_changes" = true ]; then
+		git push origin master
+	fi
 	logSuccess "Feature fix added and committed"
-	
+
 	help
 }
 
 function reset() {
 	git reset --hard $INITIAL_GIT_TAG
 	git commit -m "Resetting post demo"
-	git push origin master --force
+	if [ "$push_changes" = true ]; then
+		git push origin master --force
+	fi
 	logSuccess "Done with branch reset"
-	
+
 	help
 }
 
 logSuccess () {
-	logCustom 2 "$1"	
+	logCustom 2 "$1"
 }
 
 logInfo () {
@@ -72,8 +87,7 @@ logInfo () {
 logCustom () {
 	tput setaf $1
 	echo "$2"
-	tput sgr 0	
+	tput sgr 0
 }
 
-help				
-
+help
